Setup: A common widget for onboarding screen
