+++ lib/features/application/my_glint_app.dart
@@ -11,7 +11,7 @@ class MyGlintApp extends StatelessWidget {
       title: "Glint People App",
       debugShowCheckedModeBanner: false,
       theme: AppTheme.theme,
-      home: const OnboardingScreen(),
+      home: const WhoCatchesYourEyeOnboardingScreen(),
       // routerConfig: glintMainRoutes,
     );
   }
