+++ lib/features/exports.dart
@@ -3,4 +3,4 @@ export 'filter/filter_preference_screen.dart';
 export 'home/home_screen.dart';
 export 'likes/likes_screen.dart';
 export 'notifications/notification_screen.dart';
-export 'onboarding/onboarding_screen.dart';
+export 'onboarding/who_catches_your_eye_onboarding_screen.dart';
